+++ CHANGES.md
@@ -7,6 +7,7 @@ CHANGELOG
 
 * Add all geometries description in feature detail serializer
 * Change generated document name with feature title
+* Custom serializer for feature extra geometries
 
 
 0.3.33          (2020-02-14)